품목별 입출고 / 재고이력

+++ FamTec/FamTec/FamTec/Server/Controllers/Building/Group/BuildingGroupController.cs
@@ -1,4 +1,5 @@
-﻿using FamTec.Server.Services.Building.Group;
+﻿using FamTec.Server.Services;
+using FamTec.Server.Services.Building.Group;
 using FamTec.Shared.Server.DTO;
 using FamTec.Shared.Server.DTO.Building.Group;
 using Microsoft.AspNetCore.Authorization;
@@ -11,10 +12,13 @@ namespace FamTec.Server.Controllers.Building.Group
     public class BuildingGroupController : ControllerBase
     {
         private IBuildingGroupService GroupService;
+        private ILogService LogService;
 
-        public BuildingGroupController(IBuildingGroupService _groupservice)
+        public BuildingGroupController(IBuildingGroupService _groupservice,
+            ILogService _logservice)
         {
             this.GroupService = _groupservice;
+            this.LogService = _logservice;
         }
 
 
@@ -23,52 +27,60 @@ namespace FamTec.Server.Controllers.Building.Group
         [Route("sign/AddBuildingGroup")]
         public async ValueTask<IActionResult> AddBuildingGroup([FromBody] AddGroupDTO? dto)
         {
-            //AddGroupDTO dto = new AddGroupDTO();
-            //dto.BuildingIdx = 2;
-            //dto.Name = "옥상주차장";
-
-            //AddGroupItemKeyDTO key = new AddGroupItemKeyDTO();
-            //key.Name = "전기차";
-
-            //key.ItemValues.Add(new AddGroupItemValueDTO()
-            //{
-            //    Values = "3",
-            //    Unit = "대"
-            //});
-
-            //dto.AddGroupKey.Add(key);
-
-            //key = new AddGroupItemKeyDTO();
-            //key.Name = "수소차";
-
-            //key.ItemValues.Add(new AddGroupItemValueDTO()
-            //{
-            //    Values = "5",
-            //    Unit = "대"
-            //});
-            //key.ItemValues.Add(new AddGroupItemValueDTO()
-            //{
-            //    Values = "8",
-            //    Unit = "대"
-            //});
-            //dto.AddGroupKey.Add(key);
-
-            ResponseUnit<AddGroupDTO?> model = await GroupService.AddBuildingGroupService(HttpContext, dto);
-
-            if (model is not null)
+            try
             {
-                if (model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
+
+
+
+                //AddGroupDTO dto = new AddGroupDTO();
+                //dto.BuildingIdx = 2;
+                //dto.Name = "옥상주차장";
+
+                //AddGroupItemKeyDTO key = new AddGroupItemKeyDTO();
+                //key.Name = "전기차";
+
+                //key.ItemValues.Add(new AddGroupItemValueDTO()
+                //{
+                //    Values = "3",
+                //    Unit = "대"
+                //});
+
+                //dto.AddGroupKey.Add(key);
+
+                //key = new AddGroupItemKeyDTO();
+                //key.Name = "수소차";
+
+                //key.ItemValues.Add(new AddGroupItemValueDTO()
+                //{
+                //    Values = "5",
+                //    Unit = "대"
+                //});
+                //key.ItemValues.Add(new AddGroupItemValueDTO()
+                //{
+                //    Values = "8",
+                //    Unit = "대"
+                //});
+                //dto.AddGroupKey.Add(key);
+
+                if (HttpContext is null)
                     return BadRequest();
-                }
+                if (dto is null)
+                    return BadRequest();
+
+                ResponseUnit<AddGroupDTO?> model = await GroupService.AddBuildingGroupService(HttpContext, dto);
+
+                if (model is null)
+                    return BadRequest();
+
+                if (model.code == 200)
+                    return Ok(model);
+                else
+                    return BadRequest();
             }
-            else
+            catch(Exception ex)
             {
-                return BadRequest();
+                LogService.LogMessage(ex.Message);
+                return Problem("서버에서 처리할수 없는 작업입니다.", statusCode: 500);
             }
         }
 
@@ -77,21 +89,25 @@ namespace FamTec.Server.Controllers.Building.Group
         [Route("sign/GetBuildingGroup")]
         public async ValueTask<IActionResult> GetDetailBuilding(int? buildingid)
         {
-            ResponseList<GroupListDTO?> model = await GroupService.GetBuildingGroupListService(HttpContext, buildingid);
-            if (model is not null)
+            try
             {
-                if (model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
+                if (HttpContext is null)
                     return BadRequest();
-                }
+
+                ResponseList<GroupListDTO?> model = await GroupService.GetBuildingGroupListService(HttpContext, buildingid);
+                
+                if (model is null)
+                    return BadRequest();
+                
+                if (model.code == 200)
+                    return Ok(model);
+                else
+                    return BadRequest();
             }
-            else
+            catch (Exception ex)
             {
-                return BadRequest();
+                LogService.LogMessage(ex.Message);
+                return Problem("서버에서 처리할수 없는 작업입니다.", statusCode: 500);
             }
         }
 
@@ -100,21 +116,27 @@ namespace FamTec.Server.Controllers.Building.Group
         [Route("sign/UpdateGroup")]
         public async ValueTask<IActionResult> UpdateBuildingGroup([FromBody] UpdateGroupDTO? dto)
         {
-            ResponseUnit<bool?> model = await GroupService.UpdateGroupNameService(HttpContext, dto);
-            if (model is not null)
+            try
             {
-                if (model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
+                if (HttpContext is null)
                     return BadRequest();
-                }
+                if (dto is null)
+                    return BadRequest();
+
+                ResponseUnit<bool?> model = await GroupService.UpdateGroupNameService(HttpContext, dto);
+                
+                if (model is null)
+                    return BadRequest();
+
+                if (model.code == 200)
+                    return Ok(model);
+                else
+                    return BadRequest();
             }
-            else
+            catch(Exception ex)
             {
-                return BadRequest();
+                LogService.LogMessage(ex.Message);
+                return Problem("서버에서 처리할수 없는 작업입니다.", statusCode: 500);
             }
         }
 
@@ -123,21 +145,27 @@ namespace FamTec.Server.Controllers.Building.Group
         [Route("sign/DeleteGroup")]
         public async ValueTask<IActionResult> DeleteBuildingGroup([FromBody]int? groupid)
         {
-            ResponseUnit<bool?> model = await GroupService.DeleteGroupService(HttpContext, groupid);
-            if (model is not null)
+            try
             {
-                if (model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
+                if (HttpContext is null)
                     return BadRequest();
-                }
+                
+                if (groupid is null)
+                    return BadRequest();
+
+                ResponseUnit<bool?> model = await GroupService.DeleteGroupService(HttpContext, groupid);
+                if (model is null)
+                    return BadRequest();
+
+                if (model.code == 200)
+                    return Ok(model);
+                else
+                    return BadRequest();
             }
-            else
+            catch(Exception ex)
             {
-                return BadRequest();
+                LogService.LogMessage(ex.Message);
+                return Problem("서버에서 처리할수 없는 작업입니다.", statusCode: 500);
             }
         }
